Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15524 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/pvcamSrc/pvCam.cpp
@@ -49,7 +49,7 @@ pvCam::pvCam(const char *portName, int maxSizeX, int maxSizeY, NDDataType_t data
     const char *functionName = "pvCam::pvCam()";
     int status = asynSuccess;
     int addr=0;
-    int dims[2];
+    size_t dims[2];
 
     //init some variables
     detectorSelected = 0;
